made adjustments to root scripts in packageJSON and installed gh-pages
