+++ client/src/components/Carousel_Item/index.js
@@ -23,7 +23,7 @@ function Carousel_Item(props) {
                     </div>
                     
                     <div className="card-action">
-                        <a className="moveIcon btn-floating btn waves-effect waves-light light-blue "><Link to='Study'><i className="material-icons">menu_book</i></Link></a>
+                        <a className="moveIcon btn-floating btn waves-effect waves-light light-blue "><Link to='/study'><i className="material-icons">menu_book</i></Link></a>
                         <div class="divider"/>
                         <a className="btn-floating btn- waves-effect waves-light red"><i className="material-icons">delete_forever</i></a>
                         <div class="divider"/>
