+++ package.json
@@ -8,7 +8,10 @@
     "develop": "concurrently \"cd server && npm run watch\" \"cd client && npm start\"",
     "install": "cd server && npm i && cd ../client && npm i",
     "seed": "cd server && npm run seed",
-    "build": "cd client && npm run build"
+    "predeploy": "npm run build",
+    "deploy": "gh-pages -b main -d build",
+    "start": "react-scripts start",
+    "build": "react-scripts build"
   },
   "repository": {
     "type": "git",
@@ -20,7 +23,7 @@
   "bugs": {
     "url": "https://github.com/clabel95/Decked-Out/issues"
   },
-  "homepage": "https://github.com/clabel95/Decked-Out#readme",
+  "homepage": "https://clabel95.github.io/Decked-Out",
   "devDependencies": {
     "concurrently": "^7.0.0"
   },
@@ -30,6 +33,7 @@
     "axios": "^0.26.1",
     "bcrypt": "^5.0.1",
     "express": "^4.17.3",
+    "gh-pages": "^3.2.3",
     "graphql": "^16.3.0",
     "jwt-decode": "^3.1.2",
     "materialize-css": "^1.0.0",
